Clean up code with imports
Alas I am in crisis. I had sworn to the gods of fully qualified
imports that I would keep my imports clean. But the type annotations
become really difficult to read (and not to mention a lot to type).

I decided to import types, but will continue to use full module paths
for function calls.