Adjust serve x position for fair serves
The ball is re-served in the direction of the player that gets scored
on. But the speed of the ball mixed with the randomness of the position
and Y velocity made it too easy to get repeatedly scored on.

This pushes the starting point for serves 100 pixels to the right when
serving left, and 100 pixels to the left when serving right. The result
gives players more time to adjust to getting scored on and return new
serves.