Handle paddle hits, WIP smoothing out window hits
The code is getting a bit messy here, but I can probably take another
pass at refactoring it. Splitting the handling of paddle and wall
collisions would probably be a good idea.

I was trying to remove the manual ball repositioning in the `updateBall`
function because it _seems_ like it should work to just negate the
velocity and have the ball head back in the direction it came from. BUT
without the position adjustment the ball "sticks" to the wall and stops
moving.

The repositioning works to allow the ball to keep moving, but was adding
a janky collision against the walls. This commit fixes the issue in
terms of making for a smooth collision with the bottom edge of the
window, but there's likely a much smoother way to handle this.