Add ability to change ball angle off the paddle
This is a little messy, but it adds to the sense of control for the
player by allowing the paddle to change the direction of the ball. If it
hits slightly to the right of the center of the paddle, the ball angle
will be slightly to the right as well. Hits closer to the edge of the
paddle will cause a sharp angle, sending the ball quickly in a
horizontal path.

This also adds a slight speed change every time the ball hits the
paddle. Perhaps these settings should all be part of a global config
that's documented at the top of the file.