Experimenting with level design
Wondering if I could create a level designer where I can just toggle
on/off for blocks in a screen, and then have it output a giant list of
tuples like this.

This is one of the few times I wish there was an opt-out for elm-format
so the tuples would match the layout of the screen.