Building map for first Adventure screen
The source for the screen is in `assets/elm/static/screens` as a raw
text file. This is so `elm-format` doesn't format it. It's not ideal to
have to edit the map and then pull it into the Elm app, but the next
step would potentially be to create a Level editor that allows for
clicking on the SVG to flag certain positions in the screen.